requirements.txt

python-dotenv==1.1.0